requirements.txt
pyzbar~=0.1.9
Pillow>=10.2.0